requirements.txt
flake8
requests
pyyaml